+++ include/mnv_table.h
@@ -9,7 +9,6 @@ extern "C"
 
 
 #include <stddef.h>     // size_t
-#include <stdint.h>     // uint32_t
 
 
 /**
@@ -58,10 +57,10 @@ int
 vrd_mnv_table_insert(vrd_MNV_Table* const table,
                      size_t const len,
                      char const reference[len],
-                     uint32_t const start,
-                     uint32_t const end,
-                     uint32_t const sample_id,
-                     uint32_t const phase,
+                     size_t const start,
+                     size_t const end,
+                     size_t const sample_id,
+                     size_t const phase,
                      void* const inserted);
